Move attributes operation route factory into sylius resource namespace
