+++ src/Bundle/Routing/RoutesAttributesLoader.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Sylius\Bundle\ResourceBundle\Routing;
 
 use Sylius\Resource\Reflection\ClassReflection;
+use Sylius\Resource\Symfony\Routing\Factory\AttributesOperationRouteFactoryInterface;
 use Symfony\Bundle\FrameworkBundle\Routing\RouteLoaderInterface;
 use Symfony\Component\Routing\RouteCollection;
